pyproject.toml
[project]
name = "resume-llm-analyzer"
version = "0.1.0"
description = "AI-powered resume coach with local and cloud LLM inference"
requires-python = ">=3.10"

dependencies = [
    # API backend
    "fastapi>=0.100.0",
    "uvicorn>=0.22.0",
    "pydantic>=2.0.0",

    # Frontend
    "streamlit>=1.32.0",

    # Inference Check — Nebius AI (OpenAI-compatible) + Modal serverless GPU
    "openai>=1.30.0",
    "modal>=0.62.0",

    # Utilities
    "PyPDF2>=3.0.0",
    "boto3>=1.28.0",
    "sagemaker>=2.200.0",
    "requests>=2.31.0",
    "beautifulsoup4>=4.12.0",
    "lxml>=4.9.0",
    "python-dotenv>=1.0.0",
    "huggingface-hub>=0.18.0",
    "sentencepiece>=0.1.99",
    "protobuf>=4.24.0",
]

[project.optional-dependencies]
# Install on Linux / macOS ARM (Apple Silicon) GPU servers where torch runs natively.
# Not available on macOS x86_64 (Intel Mac) with Python 3.12+.
# Usage:  uv sync --extra local-model
local-model = [
    "torch>=2.0.0",
    "transformers>=4.30.0",
    "accelerate>=0.20.0",
    "bitsandbytes>=0.40.0",
]

[dependency-groups]
dev = [
    "pytest>=7.0.0",
    "ruff>=0.4.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["backend", "frontend", "inference_check"]

[tool.ruff]
line-length = 100
target-version = "py310"
exclude = [
    "backend/FT_llm/",       # experimental fine-tuning scripts, not production
    "backend/test_*.py",     # ad-hoc test scripts
]
